TXL-231: Fixed column sorting problems.
(Note: the sort glyph doesn't display for the ref column. There appears to be enough room for it, but I found a person who was reporting a similar problem and maybe in some cases Microsoft is off by one pixel. Probably fixable programmatically, but this may not be a regression.)